+++ CHANGELOG.md
@@ -37,6 +37,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## Unreleased
 
+### Improvements
+
+* [#725](https://github.com/babylonlabs-io/finality-provider/pull/725) chore: adds check for pk from bytes
+
 ## v2.0.0-rc.5
 
 ### Improvements